tools/editor_trends/wikilytics/requirements.txt
djangotoolbox
django-nonrel
django-mongodb-engine
